refactor(handlers): divide mock_configuration and mock_handlers
